refactor : split code
